deck: move the URL machinery out of capabilities
The projection and the navigate helpers are not capability modules, so
capabilities/ no longer holds them. They live in url/ alongside the deck write
path, which was likewise sitting in operations/ without being an operation
handler. Dependencies now run one way: operations reach into url, never back.

Navigation is exported from the util barrel rather than reached by a deep path,
and the serializer Navigation.format replaced is deleted.

Reverts the Operation and Process doc notes, which belong to a different change.
