+++ packages/plugins/plugin-deck/src/types/DeckSchema.ts
@@ -11,6 +11,8 @@ import * as Translations from '@dxos/app-toolkit/Translations';
 
 import { meta } from '#meta';
 
+import { Navigation } from '../util';
+
 export const PLANK_COMPANION_TYPE = AppNode.PLANK_COMPANION_TYPE;
 export const DECK_COMPANION_TYPE = AppNode.DECK_COMPANION_TYPE;
