+++ packages/plugins/plugin-deck/src/url/index.ts
@@ -1,0 +1,7 @@
+//
+// Copyright 2026 DXOS.org
+//
+
+export * from './apply';
+export * from './navigate';
+export * from './project';
