+++ packages/plugins/plugin-deck/src/util/index.ts
@@ -2,12 +2,12 @@
 // Copyright 2024 DXOS.org
 //
 
+export * as Navigation from './navigation';
 export * from './companion-anchor';
 export * from './layoutAppliesTopbar';
 export * from './migrate-persisted-state';
 export * from './navigation-target';
 export * from './resolve-sidebar-state';
-export * from './serialize-deck-url';
 export * from './set-active';
 export * from './upsert-toast';
 export * from './openable-children';
